chore(api/v1): Move registry package. Again ;)
